requirements.txt
absl-py==2.1.0
cachetools==4.2.4
certifi==2024.2.2
cffi==1.15.1
charset-normalizer==3.3.2
cloudpickle==1.2.2
filelock==3.0.10
future==1.0.0
glcontext==2.5.0
glfw==1.12.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
grpcio==1.62.1
gym==0.15.3
gym3==0.3.3
idna==3.7
imageio==2.31.2
imageio-ffmpeg==0.3.0
importlib-metadata==6.7.0
Markdown==3.4.4
MarkupSafe==2.1.5
moderngl==5.10.0
numpy==1.21.6
# nvidia-cublas-cu11==11.10.3.66
# nvidia-cuda-nvrtc-cu11==11.7.99
# nvidia-cuda-runtime-cu11==11.7.99
# nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
pandas==1.3.5
Pillow==9.5.0
protobuf==4.24.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyglet==1.3.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
scipy==1.7.3
six==1.16.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
typing_extensions==4.7.1
urllib3==2.0.7
Werkzeug==2.2.3
zipp==3.15.0
